pyproject.toml
[tool.poetry]
name = "dbt"
version = "0.1.0"
description = ""
authors = ["Benedikt Schröter <42349995+BenSchr@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dbt-core = "^1.3.2"
dbt-postgres = "^1.3.2"
jupyter = "^1.0.0"
psycopg2-binary = "^2.9.5"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
